test(wms): write a Worker mock for running tests with jest
Jest relies on Node which does not provide a Worker API.

In the `test-setup.js` file we define a mock of the Worker API, largely
inspired by https://github.com/developit/jsdom-worker.

This mock takes in the worker file path and uses esbuild to bundle it
into node-compatible code. This code is then used to construct a function
which replicates the Worker API (postMessage, terminate, etc.) in a
synchronous way.

Node that a transformer was added as well to make sure that the Worker
constructor received the full path to the worker file instead of a relative one
(like `new Worker('./worker.js')`).

Lastly, a custom mock of the Fetch API is defined in `test-setup.js` to allow
modifying the fetch response in the tests.