feat(cache): purge expired entries regularly, do not fail when cache is full
The local storage usually caps at 5MB, which can definitely be reached with an
expiry duration of 1h and many Getcapabilities requests.

If the cache storage is full, we just skip caching altogether. The next step
would be to switch to a different caching strategy, either using the Cache API
or indexed db