feat: add logic for trying several encodings when receiving a document
The following encodings will be attempted:
* utf-8
* utf-16le
* iso-8859-1

These should catch most cases, but a fallback to utf-8 (which may be completely
garbled) is used if none matches.

If the encoding is specified in the Content-Type header, it will be used
in priority (and if it fails, the logic above is used).