Expose provider information from capabilities
It is often required as part of an application to show the metadata
regarding the service provider which is included in capabilities
documents (WFS 1.1+, WMS 1.1+, WMTS). This commit exposes this
information in a common format as part of the GenericEndpointInfo
structure.

Note that this information is already included in the
OgcApiCollectionInfo structure.