Include scale denominator in WMS layer information
WMS 1.1.1 ScaleHint is converted to scale denominators as used
in WMS 1.3.0 capabilities documents for consistency.

Note that the test using WMS 1.1.1 currently fails due to a
significant difference in the scale hints provided in the
1.1.1 and 1.3.0 test documents, presumably caused by a
non-standard pixel size being used to calculate the scale
hints for the 1.1.1 document (i.e. not exactly a square
0.28x0.28mm pixel as specified by the WMS 1.3.0 standard).