feature: Better exports
This has the aim of exposing the existing API without using a single
barrel file, so that e.g. the http-utils functions can be sensibly used
by other packages or so that a bundler-friendly import of individual
modules can be used.

The existing exports as defined in the base index.js have been
modularised into indidual files under api/. Everything is still
re-exported in index.js ensuring the existing exported API is still
available without change, but adding the new individual exports.